Add files via upload
This project aims to explore classification models using Decision Tree and Random Forest based on the available data, as well as to predict the outcome of home loan approvals obtained from these models. This project is supported by a problem statement and data sourced from Kaggle. Furthermore, a program design will be developed and implemented using Google Colaboratory to achieve the objectives and benefits of this project. Through data analysis and model development using algorithms, it is expected to determine loan eligibility more efficiently and accurately.